requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
opencv-python==4.8.0.76
scikit-learn==1.3.0
matplotlib==3.7.2
scikit-image==0.21.0
Werkzeug==2.3.7
joblib==1.3.2
google-generativeai>=0.3.1,<0.4.0
python-dotenv>=1.0.1
# File yaml
pyyaml==6.0.2
seaborn==0.13.2
yellowbrick==1.5

